Enabled Test Outdated card for RA Tests on homescreen added check for outdatedness to RA getState Deletion button currently shows no warining dialog

+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/appconfig/CoronaTestConfig.kt
@@ -9,10 +9,10 @@ interface CoronaTestConfig {
 }
 
 data class CoronaRapidAntigenTestParametersContainer(
-    val hoursToDeemTestOutdated: Int = DEFAULT_HOURS
+    val hoursToDeemTestOutdated: Long = DEFAULT_HOURS
 ) {
     companion object {
-        const val DEFAULT_HOURS = 48
+        const val DEFAULT_HOURS = 48L
     }
 }
